# Conflicts:
#	todo/bin/main/com/example/todo/Controller/HomeController.class
#	todo/bin/main/static/css/common.css
#	todo/bin/main/templates/index.html
#	todo/src/main/java/com/example/todo/Controller/HomeController.java
#	todo/src/main/resources/static/css/common.css
#	todo/src/main/resources/templates/index.html

+++ todo/src/main/java/com/example/todo/Model/Todo.java
@@ -23,6 +23,6 @@ public class Todo {
 	private int goalCount;
 	private int initCount;
 	private String range;
-	private int party_ID;
+	private long party_ID;
 	
 }
